Handle any exception that gets thrown during JWT verification
The seed project's CLIENT_SECRET is the app's client secret.
However, in the case of using RS256, this is the wrong secret. So,
when verification occurs, a DomainException is thrown from the
openssl code as the secret is not a valid public key. So, at least
handle the exception and return a 500 so the frontend can treat it
as an error, rather than a 200 OK when it didn't actually work
(I've submitted a pull-request to handle the error through the
angular seed project too)