Composer/PHPCS: use a Composer plugin to sort out the PHPCS installed standards
For PHP_CodeSniffer to recognize external standards, the `installed_paths` command needs to be run.
The DealerDirect Composer PHPCS plugin can sort this out automatically for you, no matter which or how many external standards would be in use.

This plugin makes sure that the `PHPCompatibility` standard which is being installed via Composer can actually be used.

The relevant command will automatically be run after each `composer install/update`.

For convenience, I've added an additional script to the `composer.json` file to allow for calling the relevant command on demand.

References:
* https://github.com/squizlabs/PHP_CodeSniffer/wiki/Configuration-Options#setting-the-installed-standard-paths
* https://github.com/DealerDirect/phpcodesniffer-composer-installer