Update callback verification in HttpRequest.  (#769)
### Changes

'callback' is a property not a method of $mockedResponse

Changed the method_exists check to a property_exists check in the
HttpRequest class. This adjustment ensures that 'callback' is an
accessible property of the object before attempting to invoke it. This
change improves code reliability, specifically for the cases where
HttpRequest is used in unit tests with mock responses.

### References

[#768 Callback Never
Called](https://github.com/auth0/auth0-PHP/issues/768)

### Testing

A test has been added to assert the fix is working as expected

### Contributor Checklist

- [x] I agree to adhere to the [Auth0 General Contribution
Guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md).
- [x] I agree to uphold the [Auth0 Code of
Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md).